Move custom autocast modules to separate files in a custom_modules/ directory.
